Introduce StyledElement base class
